requirements.txt
alabaster==0.7.12
anytree==2.6.0
apipkg==1.5
appdirs==1.4.3
arrow==0.13.0
asn1crypto==0.24.0
astroid==2.1.0
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.3
Babel==2.6.0
bcrypt==3.1.6
Beaker==1.10.1
beautifulsoup4==4.6.0
binaryornot==0.4.4
black==19.3b0
blis==0.1.0
bs4==0.0.1
btrfsutil==1.1.0
CacheControl==0.12.5
cached-property==1.5.1
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
cookiecutter==1.6.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
coverage==4.5.3
cryptography==2.6.1
cycler==0.10.0
cymem==2.0.2
Cython==0.29.6
cytoolz==0.9.0
dill==0.2.8.2
distlib==0.2.8
distro==1.4.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.5
django-ckeditor==5.6.1
django-debug-toolbar==1.11
django-extensions==2.1.4
django-js-asset==1.1.0
django-modelcluster==4.3
django-taggit==0.23.0
django-toolbelt==0.0.1
django-treebeard==4.3
django-widget-tweaks==1.4.3
djangorestframework==3.9.1
docker==3.7.2
docker-compose==1.24.0
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
draftjs-exporter==2.1.5
en-core-web-sm==2.0.0
future==0.17.1
gunicorn==19.9.0
html5lib==1.0.1
httpie==1.0.2
hypothesis==4.14.5
idna==2.8
imagesize==1.1.0
iniconfig==1.0.0
isort==4.3.4
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10
jinja2-time==0.2.0
jsonschema==2.6.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lensfun==0.3.2
lockfile==0.12.2
lxml==4.3.3
Mako==1.0.8
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mccabe==0.6.1
meson==0.49.2
mock==2.0.0
more-itertools==4.3.0
msgpack==0.6.1
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
mysql==0.0.2
mysqlclient==1.4.2
numpy==1.16.2
packaging==19.0
paramiko==2.4.2
parso==0.3.4
pbr==5.1.3
pep517==0.5.0
Pillow==5.4.1
pipenv==2018.11.26
plac==0.9.6
pluggy==0.9.0
ply==3.11
poyo==0.4.2
preshed==2.0.1
progress==1.5
psycopg2==2.7.7
psycopg2-binary==2.7.7
pt-core-news-sm==2.0.0
pwquality==1.4.0
py==1.8.0
pyalsa==1.1.6
pyasn1==0.4.5
pycairo==1.18.0
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.3.1
PyGObject==3.32.0
pylast==3.1.0
pylint==2.2.2
pylint-django==2.0.6
pylint-plugin-utils==0.5
pyls-black==0.4.0
PyNaCl==1.3.0
pyparsing==2.3.1
pytest==4.4.0
python-dateutil==2.8.0
python-jsonrpc-server==0.1.2
python-language-server==0.25.0
python-libtorrent==1.1.12
pytoml==0.1.20
pytz==2018.9
PyYAML==3.13
Reflector==2019.3.8.1.54.39
regex==2018.1.10
requests==2.21.0
retrying==1.3.3
six==1.12.0
snowballstemmer==1.2.1
soupsieve==1.9
spacy==2.0.18
speedtest-cli==2.1.1
Sphinx==1.8.4
sphinxcontrib-websupport==1.1.0
sqlparse==0.2.4
static3==0.7.0
team==1.0
texttable==1.6.1
thinc==6.12.1
toml==0.10.0
toolz==0.9.0
tqdm==4.28.1
typed-ast==1.3.1
ujson==1.35
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.2.0
virtualenv-clone==0.4.0
wagtail==2.4
wagtailcodeblock==1.14.0.0
webencodings==0.5.1
websocket-client==0.56.0
whichcraft==0.5.2
whitenoise==4.1.2
Willow==1.1
wrapt==1.11.1
